pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "himena-image"
dynamic = ["version"]
description = 'Himena plugin for image processing and analysis'
readme = "README.md"
requires-python = ">=3.9"
license = { file = "LICENSE" }
keywords = []
authors = [
  { name = "hanjinliu", email = "liuhanjin.sc@gmail.com" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
  "himena",
  "ndv",
  "impy-array>=2.4.3",
]

[project.urls]
Documentation = "https://github.com/hanjinliu/himena-image#readme"
Issues = "https://github.com/hanjinliu/himena-image/issues"
Source = "https://github.com/hanjinliu/himena-image"

[tool.hatch.version]
path = "src/himena_image/__init__.py"

[tool.hatch.envs.types]
extra-dependencies = [
  "mypy>=1.0.0",
]
[tool.hatch.envs.types.scripts]
check = "mypy --install-types --non-interactive {args:src/himena_image tests}"

[tool.coverage.run]
source_pkgs = ["himena_image"]
branch = true
parallel = true
omit = [
  "src/himena_image/__init__.py",
]

[tool.coverage.paths]
himena_image = ["src/himena_image", "*/himena-image/src/himena_image"]
tests = ["tests", "*/himena-image/tests"]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]
